Add further aspects
All added aspects are of course put up to discussion. We should probably split the "General" section in further subsections, because it seems to have grown to a "Miscellaneous" section. Same holds for "Development". In my opinion there should be a section about "Security", too -- I tried to imply it with "security by design". Maybe we could rearrange the whole section according to DIN ISO 9126?